requirements.txt
absl-py==2.0.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.2.0
deepface==0.0.79
fastapi==0.103.2
fire==0.5.0
flatbuffers==23.5.26
gast==0.5.4
gdown==4.7.1
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-transliteration-api==1.0.3
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
libclang==16.0.6
ml-dtypes==0.2.0
mtcnn==0.1.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
protobuf==4.24.3
pydantic==2.4.2
pydantic_core==2.10.1
pyreadline3==3.4.1
python-multipart==0.0.6
pyttsx3==2.90
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
termcolor==2.3.0
uvicorn[standard]
